backend/requirements.txt
uvicorn
pydantic
groq
duckduckgo-search
streamlit
tavily-python